#52 Move structg implementation into compact subpackage
