+++ compact/structg/doc.go
