+++ compact/structg/new.go
