+++ compact/structg/node.go
