+++ compact/structg/tree.go
