+++ compact/structg/tree_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 
 	"github.com/alex-ilchukov/radixt"
+	"github.com/alex-ilchukov/radixt/compact/structg"
 	"github.com/alex-ilchukov/radixt/generic"
-	"github.com/alex-ilchukov/radixt/structg"
 )
 
 var (
